executor/requirements.txt
minio
psycopg2-binary
python-dotenv